==> q1_submission_notebooks/11_database_searcher.ipynb <==
import numpy as np
import pandas as pd

df = pd.read_csv('data/Perovskite_database_content_all_data.csv')

relevant = df[['Stability_protocol', 'JV_default_Voc', 'Stability_relative_humidity_average_value', 'Stability_time_total_exposure', 'Stabilised_performance_PCE', 'Stability_PCE_end_of_experiment', 'Stability_PCE_initial_value', 'Stability_temperature_range', 'Perovskite_composition_long_form', 'Ref_publication_date', 'Ref_journal', 'Cell_architecture', 'Cell_stack_sequence', 'Ref_DOI_number' ]]

def search_database(search_term):
    return relevant[relevant['Ref_DOI_number'].str.contains(search_term, case=False) == True]

print(search_database('jacs.5b11740'))

print(search_database('s41586'))
